DVC/requirement.txt
pandas
snscrape
pickle
textblob
scikit-learn
nltk
string
json
pyyaml
jasonify
